Corregimos estilos de la página. Agregamos validaciones de manera correcta, Agregamos mensajes al realizar acciones
