--- src/components/Navbar.jsx
@@ -1,13 +1,0 @@
-import React from "react";
-import { Link } from "react-router-dom";
-import "../styles/main.css"; 
-import loginIcon from "/Img/login.svg";
-import carritoIcon from "/Img/carrito.svg";
-
-const Navbar = () => {
-    return (
-        <>
-            <h1>Hola</h1>
-        </>
-    )
-}
