+++ src/pages/catalogo.jsx
@@ -1,6 +1,5 @@
 import React from "react";
 import { Link } from "react-router-dom";
-import "../styles/catalogo.css";
 
 const Catalogo = () => {
     return (
