+++ src/pages/contacto.jsx
@@ -53,89 +53,6 @@ const Contact = () => {
 
     return (
         <>
-        <div className="sticky">
-
-            {/* Navbar Mobile */}
-            <nav className="navbar-mobile">
-            <Link to="/" className="tittle">
-                NICAXEL
-            </Link>
-
-            <ul className="nav-menu">
-                <li className="nav-item">
-                <Link to="/" className="nav-link">
-                    Home
-                </Link>
-                </li>
-
-                <li className="nav-item">
-                <Link to="/nosotros" className="nav-link">
-                    Sobre nosotros
-                </Link>
-                </li>
-
-                <li className="nav-item">
-                <Link to="/" className="nav-link">
-                    Blog
-                </Link>
-                </li>
-
-                <li className="nav-item">
-                <Link to="/catalogo" className="nav-link">
-                    Catálogo
-                </Link>
-                </li>
-
-                <li className="nav-item">
-                <div className="login-flex">
-                    <Link to="/login" className="nav-link">
-                    Iniciar sesión
-                    </Link>
-                    <img src="/Img/login.svg" alt="icono login" />
-                </div>
-                </li>
-
-                <li className="nav-item">
-                <Link to="/carrito" className="nav-link">
-                    Ver carrito 🛒(2)
-                </Link>
-                </li>
-            </ul>
-
-            <div className="burger">
-                <span className="bar"></span>
-                <span className="bar"></span>
-                <span className="bar"></span>
-            </div>
-            </nav>
-
-            {/* Navbar Desktop */}
-            <nav className="navbar-dekstop">
-            <Link to="/" className="tittle">
-                NICAXEL
-            </Link>
-
-            <div className="nav-medio">
-                <Link to="/nosotros">Sobre nosotros</Link>
-                <Link to="/">Blog</Link>
-                <span>Contacto</span>
-            </div>
-
-            <div className="nav-derecha">
-                <Link to="/login">
-                <img src="/Img/login.svg" alt="Icono de login" />
-                </Link>
-                <div className="carrito-flex">
-                <Link to="/carrito">
-                    <img src="/Img/carrito.svg" alt="Icono de carrito" />
-                </Link>
-                <span>(1)</span>
-                </div>
-            </div>
-            </nav>
-        </div>
-
-        {/* Contenido principal */}
         <main>
             <h1>Contáctanos</h1>
