+++ src/pages/login.jsx
@@ -54,78 +54,6 @@ const Login = () => {
 
     return (
         <>
-        <div className="sticky">
-
-            {/* Navbar Mobile */}
-            <nav className="navbar-mobile">
-            <Link to="/" className="tittle">
-                NICAXEL
-            </Link>
-
-            <ul className="nav-menu">
-                <li className="nav-item">
-                <Link to="/" className="nav-link">
-                    Inicio
-                </Link>
-                </li>
-                <li className="nav-item">
-                <Link to="/nosotros" className="nav-link">
-                    Sobre nosotros
-                </Link>
-                </li>
-                <li className="nav-item">
-                <Link to="/catalogo" className="nav-link">
-                    Catálogo
-                </Link>
-                </li>
-                <li className="nav-item">
-                <Link to="/contacto" className="nav-link">
-                    Contacto
-                </Link>
-                </li>
-                <li className="nav-item">
-                <Link to="/register" className="nav-link">
-                    Regístrate
-                </Link>
-                </li>
-                <li className="nav-item">
-                <Link to="/carrito" className="nav-link">
-                    Ver carrito 🛒
-                </Link>
-                </li>
-            </ul>
-
-            <div className="burger">
-                <span className="bar"></span>
-                <span className="bar"></span>
-                <span className="bar"></span>
-            </div>
-            </nav>
-
-            {/* Navbar Desktop */}
-            <nav className="navbar-dekstop">
-            <Link to="/" className="tittle">
-                NICAXEL
-            </Link>
-
-            <div className="nav-medio">
-                <Link to="/nosotros">Sobre nosotros</Link>
-                <Link to="/">Blog</Link>
-                <Link to="/contacto" className="navbar-contacto">
-                Contacto
-                </Link>
-            </div>
-
-            <div className="nav-derecha">
-                <Link to="/carrito">
-                <img src="/Img/carrito.svg" alt="Icono de carrito" />
-                </Link>
-                <span>(1)</span>
-            </div>
-            </nav>
-        </div>
-
-        {/* MAIN */}
         <main>
             <h1>Regístrate o inicia sesión</h1>
