+++ src/pages/nosotros.jsx
@@ -6,87 +6,6 @@ import "../styles/nosotros.css";
 function Nosotros() {
     return (
         <>
-        <div className="sticky">
-        
-
-            {/* Navbar Mobile */}
-            <nav className="navbar-mobile">
-            <Link to="/" className="tittle">
-                NICAXEL
-            </Link>
-
-            <ul className="nav-menu">
-                <li className="nav-item">
-                <Link to="/" className="nav-link">
-                    Home
-                </Link>
-                </li>
-                <li className="nav-item">
-                <Link to="/" className="nav-link">
-                    Blog
-                </Link>
-                </li>
-                <li className="nav-item">
-                <Link to="/catalogo" className="nav-link">
-                    Catálogo
-                </Link>
-                </li>
-                <li className="nav-item">
-                <Link to="/contacto" className="nav-link">
-                    Contacto
-                </Link>
-                </li>
-                <li className="nav-item">
-                <div className="login-flex">
-                    <Link to="/login" className="nav-link">
-                    Iniciar sesión
-                    </Link>
-                    <img src="/Img/login.svg" alt="icono login" />
-                </div>
-                </li>
-                <li className="nav-item">
-                <Link to="/carrito" className="nav-link">
-                    Ver carrito 🛒(2)
-                </Link>
-                </li>
-            </ul>
-
-            <div className="burger">
-                <span className="bar"></span>
-                <span className="bar"></span>
-                <span className="bar"></span>
-            </div>
-            </nav>
-
-            {/* Navbar Desktop */}
-            <nav className="navbar-dekstop">
-            <Link to="/" className="tittle">
-                NICAXEL
-            </Link>
-
-            <div className="nav-medio">
-                <span>Sobre nosotros</span>
-                <Link to="/">Blog</Link>
-                <Link to="/contacto" className="navbar-contacto">
-                Contacto
-                </Link>
-            </div>
-
-            <div className="nav-derecha">
-                <Link to="/login">
-                <img src="/Img/login.svg" alt="Icono de login" />
-                </Link>
-                <div className="carrito-flex">
-                <Link to="/carrito">
-                    <img src="/Img/carrito.svg" alt="Icono de carrito" />
-                </Link>
-                <span>(1)</span>
-                </div>
-            </div>
-            </nav>
-        </div>
-
-        {/* Contenido principal */}
         <main className="main-faq">
             <header className="header-cont">
             <h1>Preguntas Frecuentes</h1>
