+++ src/pages/register.jsx
@@ -41,79 +41,6 @@ function Register() {
 
     return (
         <>
-        {/* Banner superior */}
-        <div className="sticky">
-
-            {/* Navbar móvil */}
-            <nav className="navbar-mobile">
-            <Link to="/" className="tittle">
-                NICAXEL
-            </Link>
-
-            <ul className="nav-menu">
-                <li className="nav-item">
-                <Link to="/" className="nav-link">
-                    Inicio
-                </Link>
-                </li>
-                <li className="nav-item">
-                <Link to="/nosotros" className="nav-link">
-                    Sobre nosotros
-                </Link>
-                </li>
-                <li className="nav-item">
-                <Link to="/catalogo" className="nav-link">
-                    Catálogo
-                </Link>
-                </li>
-                <li className="nav-item">
-                <Link to="/contacto" className="nav-link">
-                    Contacto
-                </Link>
-                </li>
-                <li className="nav-item">
-                <Link to="/login" className="nav-link">
-                    Iniciar sesión
-                </Link>
-                </li>
-                <li className="nav-item">
-                <Link to="/carrito" className="nav-link">
-                    Ver carrito 🛒(1)
-                </Link>
-                </li>
-            </ul>
-
-            <div className="burger">
-                <span className="bar"></span>
-                <span className="bar"></span>
-                <span className="bar"></span>
-            </div>
-            </nav>
-
-            {/* Navbar escritorio */}
-            <nav className="navbar-dekstop">
-            <Link to="/" className="tittle">
-                NICAXEL
-            </Link>
-
-            <div className="nav-medio">
-                <Link to="/nosotros">Sobre nosotros</Link>
-                <Link to="/">Blog</Link>
-                <Link to="/contacto" className="navbar-contacto">
-                Contacto
-                </Link>
-            </div>
-
-            <div className="nav-derecha">
-                <Link to="/carrito">
-                <img src="/Img/carrito.svg" alt="Icono de carrito" />
-                </Link>
-                <span>(1)</span>
-            </div>
-            </nav>
-        </div>
-
-        {/* Contenido principal */}
         <main>
             <h1>Regístrate o inicia sesión</h1>
