+++ src/styles/contacto.css
@@ -1,26 +1,18 @@
-nav .nav-medio span {
-    border-bottom: 2px solid black;
-}
-
 main {
     display: grid;
-}
-
-@view-transition {
-    navigation: auto;
+    background: white;
 }
 
 h1 {
     text-align: center;
     font-size: 3.5rem;
-    margin-top: 4rem;
+    margin-top: 40px;
     margin-bottom: 2rem;
 }
 
 .formulario-div {
     display: grid;
     margin: 1rem auto;
-    max-width: 50rem;
     width: 90%;
     border-radius: .5rem;
     border: 1px solid rgb(226, 226, 226);
@@ -34,7 +26,6 @@ h1 {
 
 form {
     display: grid;
-    align-content: start;
     gap: .8rem;
     padding: 2rem 1.7rem;
 }
@@ -97,6 +88,12 @@ footer {
 }
 
 #contador{
-    margin-top: -.8rem;
+    margin-top: -.5rem;
     font-size: 1.1rem;
+}
+
+@media screen and (width > 700px) {
+h1 {
+    margin-top: -60px;
+}
 }
