+++ src/styles/main.css
@@ -4,12 +4,12 @@
     box-sizing: border-box;
 }
 
-body {
-    font-family: "Raleway", sans-serif;
-}
-
 html {
     font-size: 62.5%;
+}
+
+body {
+    font-family: "Raleway", sans-serif;
 }
 
 a {
@@ -18,10 +18,6 @@ a {
 
 ul li {
     list-style: none;
-}
-
-@view-transition {
-    navigation: auto;
 }
 
 /* ---------------------- NAVBAR ------------------------- */
@@ -538,6 +534,7 @@ footer {
     }
 
     .header-contenido {
+        padding-right: 20rem;
         padding-right: 20rem;
     }
